% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UScensuscounty20.R
\docType{data}
\name{UScensuscounty20}
\alias{UScensuscounty20}
\alias{alabamacounty20}
\alias{alaskacounty20}
\alias{arizonacounty20}
\alias{arkansascounty20}
\alias{californiacounty20}
\alias{coloradocounty20}
\alias{connecticutcounty20}
\alias{delawarecounty20}
\alias{dccounty20}
\alias{floridacounty20}
\alias{georgiacounty20}
\alias{hawaiicounty20}
\alias{idahocounty20}
\alias{illinoiscounty20}
\alias{indianacounty20}
\alias{iowacounty20}
\alias{kansascounty20}
\alias{kentuckycounty20}
\alias{louisianacounty20}
\alias{mainecounty20}
\alias{marylandcounty20}
\alias{massachussetscounty20}
\alias{michigancounty20}
\alias{minnesotacounty20}
\alias{mississippicounty20}
\alias{missouricounty20}
\alias{montanacounty20}
\alias{nebraskacounty20}
\alias{nevadacounty20}
\alias{newhampshirecounty20}
\alias{newjerseycounty20}
\alias{newmexicocounty20}
\alias{newyorkcounty20}
\alias{northcarolinacounty20}
\alias{northdakotacounty20}
\alias{ohiocounty20}
\alias{oklahomacounty20}
\alias{oregoncounty20}
\alias{pennsylvaniacounty20}
\alias{puertoricocounty20}
\alias{rhodeislandcounty20}
\alias{southcarolinacounty20}
\alias{southdakotacounty20}
\alias{tennessecounty20}
\alias{texascounty20}
\alias{utahcounty20}
\alias{vermontcounty20}
\alias{virginiacounty20}
\alias{washingtoncounty20}
\alias{westvirginiacounty20}
\alias{wisconsincounty20}
\alias{wyomingcounty20}
\title{UScensuscounty20
This package contains county level data for the US census 2020 redistricting files and shape files.
\describe{
\item{\code{FILEID}}{File Identification}
\item{\code{STUSAB}}{State/US-Abbreviation (USPS)}
\item{\code{SUMLEV}}{Summary Level=050 for county}
\item{\code{GEOVAR}}{Geographic Variant}
\item{\code{GEOCOMP}}{Geographic Component}
\item{\code{CHARITER}}{Characteristic Iteration}
\item{\code{CIFSN}}{Characteristic Iteration File Sequence Number}
\item{\code{LOGRECNO}}{Logical Record Number}
\item{\code{GEOID}}{Geographic Record Identifier}
\item{\code{GEOCODE}}{Geographic Code Identifier}
\item{\code{REGION}}{Region}
\item{\code{DIVISION}}{Division}
\item{\code{STATE}}{State}
\item{\code{COUNTY}}{County}
\item{\code{COUNTYCC}}{FIPS County class code}
\item{\code{COUNTYNS}}{County (NS)}
\item{\code{COUSUB}}{County Sub Divison (FIPS)}
\item{\code{COUSUBCC}}{FIPS County Subdivision Class Code}
\item{\code{COUSUBNS}}{County Subdivision (NS)}
\item{\code{SUBMCD}}{Subminor Civil Division (FIPS)}
\item{\code{SUBMCDCC}}{FIPS Subminor Civil Division Class Code}
\item{\code{SUBMCDNS}}{Subminor Civil Division (NS)}
\item{\code{ESTATE}}{Estate (FIPS)}
\item{\code{ESTATECC}}{FIPS Estate Class Code}
\item{\code{ESTATENS}}{Estate (NS)}
\item{\code{CONCIT}}{Consolidated City (FIPS)}
\item{\code{CONCITCC}}{FIPS Consolidated City Class Code}
\item{\code{CONCITNS}}{Consolidated City (NS)}
\item{\code{PLACE}}{Place (FIPS)}
\item{\code{PLACECC}}{FIPS Place Class Code}
\item{\code{PLACENS}}{Place (NS)}
\item{\code{TRACT}}{Census Tract}
\item{\code{BLKGRP}}{Block Group}
\item{\code{BLOCK}}{Block}
\item{\code{AIANHH}}{American Indian Area/Alaska Native Area/Hawaiian Home Land (Census)}
\item{\code{AIHHTLI}}{American Indian Trust Land/Hawaiian Home Land Indicator}
\item{\code{AIANHHFP}}{American Indian Area/Alaska Native Area/Hawaiian Home Land (FIPS)}
\item{\code{AIANHHCC}}{FIPS American Indian Area/Alaska Native Area/Hawaiian Home Land Class Code}
\item{\code{AIANHHNS}}{American Indian Area/Alaska Native Area/Hawaiian Home Land (NS)}
\item{\code{AITS}}{American Indian Tribal Subdivision (Census)}
\item{\code{AITSFP}}{American Indian Tribal Subdivision (FIPS)}
\item{\code{AITSCC}}{FIPS American Indian Tribal Subdivision Class Code}
\item{\code{AITSNS}}{American Indian Tribal Subdivision (NS)}
\item{\code{TTRACT}}{Tribal Census Tract}
\item{\code{TBLKGRP}}{Tribal Block Group}
\item{\code{ANRC}}{Alaska Native Regional Corporation (FIPS)}
\item{\code{ANRCCC}}{FIPS Alaska Native Regional Corporation Class Code}
\item{\code{ANRCNS}}{Alaska Native Regional Corporation (NS)}
\item{\code{CBSA}}{Metropolitan Statistical Area/Micropolitan Statistical Area}
\item{\code{MEMI}}{Metropolitan/Micropolitan Indicator}
\item{\code{CSA}}{Combine Statistical Area}
\item{\code{METDIV}}{Metropolitan Division}
\item{\code{NECTA}}{New England City and Town Area}
\item{\code{NMEMI}}{NECTA Metropolitan/Micropolitan Indicator}
\item{\code{CNECTA}}{Combined New England City and Town Area}
\item{\code{NECTADIV}}{New England city and town division}
\item{\code{CBSAPCI}}{Metropolitan Statistical Area/Micropolitan Statistical Area Principal City Indicator}
\item{\code{NECTAPCI}}{New England City and Town Area Principal City Indicator}
\item{\code{UA}}{Urban Area}
\item{\code{UATYPE}}{Urban Area Type}
\item{\code{UR}}{Urban/Rural}
\item{\code{CD116}}{Congressional District (116th)}
\item{\code{CD118}}{Congressional District (118th)}
\item{\code{CD119}}{Congressional District (119th)}
\item{\code{CD120}}{Congressional District (120th)}
\item{\code{CD121}}{Congressional District (121th)}
\item{\code{SLDU18}}{State Legislative District (Upper Chamber) (2018)}
\item{\code{SLDU22}}{State Legislative District (Upper Chamber) (2022)}
\item{\code{SLDU24}}{State Legislative District (Upper Chamber) (2024)}
\item{\code{SLDU26}}{State Legislative District (Upper Chamber) (2026)}
\item{\code{SLDU28}}{State Legislative District (Upper Chamber) (2028)}
\item{\code{SLDL18}}{State Legislative District (Lower Chamber) (2018)}
\item{\code{SLDL22}}{State Legislative District (Lower Chamber) (2022)}
\item{\code{SLDL24}}{State Legislative District (Lower Chamber) (2024)}
\item{\code{SLDL26}}{State Legislative District (Lower Chamber) (2026)}
\item{\code{SLDL28}}{State Legislative District (Lower Chamber) (2028)}
\item{\code{VTD}}{Voting District}
\item{\code{VTDI}}{Voting District Indicator}
\item{\code{ZCTA}}{ZIP Code Tabulation Area (5-Digit)}
\item{\code{SDELM}}{School District (Elementary)}
\item{\code{SDSEC}}{School District (Secondary)}
\item{\code{SDUNI}}{School District (Unifield)}
\item{\code{PUMA}}{Public Use Microdata Area}
\item{\code{AREALAND}}{Area(Land)}
\item{\code{AREAWATR}}{Area(Water)}
\item{\code{BASENAME}}{Area Base Name}
\item{\code{NAME}}{Area Name-Legal/Statistical Area Description (LSAD) Term-Part Indicator}
\item{\code{FUNCSTAT}}{Functional Status Code}
\item{\code{GCUNI}}{Geographic Change User Note Indicator}
\item{\code{POP100}}{Population Count (100\%)}
\item{\code{HU100}}{Housing Unit Count (100\%)}
\item{\code{INTPTLAT}}{Internal Point (Latitude)}
\item{\code{INTPTLON}}{Internal Point (Longitude)}
\item{\code{LSADC}}{Legal/Statistical Area Description Code}
\item{\code{PARTFLAG}}{Part Flag}
\item{\code{UGA}}{Urban Growth Area}
\item{\code{P0010001}}{Total Population}
\item{\code{P0010002}}{Population of single race}
\item{\code{P0010003}}{Population of white race alone}
\item{\code{P0010004}}{Population of Black or African american population alone}
\item{\code{P0010005}}{American Indian and Alaskan native alone}
\item{\code{P0010006}}{Asian Alone}
\item{\code{P0010007}}{Native Hawaiian and Other Pacific Islander alone}
\item{\code{P0010008}}{Some other race alone}
\item{\code{P0010009}}{Population of two or more races}
\item{\code{P0010010}}{Population of two races}
\item{\code{P0010011}}{White and black or african american population}
\item{\code{P0010012}}{White and American Indian or alaskan native population}
\item{\code{P0010013}}{White and Asian population}
\item{\code{P0010014}}{White and Native Hawaiian and other Pacific Islander population}
\item{\code{P0010015}}{White and some other race}
\item{\code{P0010016}}{Black or African American and American Indian and Alaska Native Population}
\item{\code{P0010017}}{Black or African American and Asian Population}
\item{\code{P0010018}}{Black or African American and Native Hawaiian and other Pacific Islander Population}
\item{\code{P0010019}}{Black or African American and some other race population}
\item{\code{P0010020}}{American Indian and Alaska Native and Asian Population}
\item{\code{P0010021}}{American Indian and Alaska Native and Native Hawaiian and other pacific islander population}
\item{\code{P0010022}}{American Indian and Alaska Native and some other race population}
\item{\code{P0010023}}{Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0010024}}{Asian and some other race population}
\item{\code{P0010025}}{Native Hawaiian and Other Pacific Islander; Some Other Race population }
\item{\code{P0010026}}{Population of three races}
\item{\code{P0010027}}{white; Black or African American; American Indian and Alaska Native population}
\item{\code{P0010028}}{White; Black or african american; Asian Population}
\item{\code{P0010029}}{white; Black or African American; Native Hawaiian and other pacific islander population}
\item{\code{P0010030}}{White; Black or African American; Some Other Race population}
\item{\code{P0010031}}{White;  American Indian and Alaska Native; Asian population }
\item{\code{P0010032}}{White; American Indian and Alaska Native; Native Hawaiian and other pacific islander population}
\item{\code{P0010033}}{White; American Indian and Alaska Native;  Some other race population}
\item{\code{P0010034}}{White; Asian; Native Hawaiian and other pacific Islander Population}
\item{\code{P0010035}}{White; Asian; Some other race population}
\item{\code{P0010036}}{White; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0010037}}{Black or African American; American Indian and Alaska Native; Asian Population}
\item{\code{P0010038}}{Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander population}
\item{\code{P0010039}}{Black or African American; American Indian and Alaska Native; Some Other Race population}
\item{\code{P0010040}}{Black or African American; Asian; Native Hawaiian and Other Pacific Islander Population}
\item{\code{P0010041}}{Black or African American; Asian; Some Other Race population}
\item{\code{P0010042}}{Black or African American; Native Hawaiian and Other Pacific Islander; Some Other Race population}
\item{\code{P0010043}}{American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander Population}
\item{\code{P0010044}}{American Indian and Alaska Native; Asian; Some Other Race population}
\item{\code{P0010045}}{American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race population}
\item{\code{P0010046}}{Asian; Native Hawaiian and Other Pacific Islander; Some Other Race population}
\item{\code{P0010046}}{Asian; Native Hawaiian and Other Pacific Islander; Some Other Race population}
\item{\code{P0010047}}{Population of four races}
\item{\code{P0010048}}{White; Black or African American; American Indian and Alaska Native; Asian}
\item{\code{P0010049}}{White; Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander}
\item{\code{P0010050}}{White; Black or African American; American Indian and Alaska Native; Some Other Race}
\item{\code{P0010051}}{White; Black or African American; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0010052}}{White; Black or African American; Asian; Some Other Race}
\item{\code{P0010053}}{White; Black or African American; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0010054}}{White; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0010055}}{White; American Indian and Alaska Native; Asian; Some Other Race}
\item{\code{P0010056}}{White; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0010057}}{White; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0010058}}{Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0010059}}{Black or African American; American Indian and Alaska Native; Asian; Some Other Race}
\item{\code{P0010060}}{Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0010061}}{Black or African American; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0010062}}{American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0010063}}{Population of five races}
\item{\code{P0010064}}{White; Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0010065}}{White; Black or African American; American Indian and Alaska Native; Asian; Some Other Race}
\item{\code{P0010066}}{White; Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0010067}}{White; Black or African American; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0010068}}{White; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0010069}}{Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0010070}}{Population of six races}
\item{\code{P0010071}}{White; Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0020001}}{Total:}
\item{\code{P0020002}}{Hispanic or Latino}
\item{\code{P0020003}}{Not Hispanic or Latino}
\item{\code{P0020004}}{Population of one race}
\item{\code{P0020005}}{White alone}
\item{\code{P0020006}}{Black or african american alone}
\item{\code{P0020007}}{American Indian or Alaskan Native alone}
\item{\code{P0020008}}{Asian Alone}
\item{\code{P0020009}}{Native Hawaiian and Other Pacific Islander alone}
\item{\code{P0020010}}{Some other race alone}
\item{\code{P0020011}}{Population of two or more races}
\item{\code{P0020012}}{Population of two races}
\item{\code{P0020013}}{White; Black or African American}
\item{\code{P0020014}}{White; American Indian and Alaska Native}
\item{\code{P0020015}}{White; Asian}
\item{\code{P0020016}}{White; Native Hawaiian and Other Pacific Islander}
\item{\code{P0020017}}{White; Some other race}
\item{\code{P0020018}}{Black or African American; American Indian and Alaska Native}
\item{\code{P0020019}}{Black or African American; Asian}
\item{\code{P0020020}}{Black or African American; Native Hawaiian and Other Pacific Islander}
\item{\code{P0020021}}{Black or African American; Some Other Race}
\item{\code{P0020022}}{American Indian and Alaska Native; Asian}
\item{\code{P0020023}}{American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander}
\item{\code{P0020024}}{American Indian and Alaska Native; Some Other Race}
\item{\code{P0020025}}{Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0020026}}{Asian; Some Other Race}
\item{\code{P0020027}}{Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0020028}}{Population of three races}
\item{\code{P0020029}}{White; Black or African American; American Indian and Alaska Native}
\item{\code{P0020030}}{White; Black or African American; Asian}
\item{\code{P0020031}}{White; Black or African American; Native Hawaiian and Other Pacific Islander}
\item{\code{P0020032}}{White; Black or African American; Some Other Race}
\item{\code{P0020033}}{White; American Indian and Alaska Native; Asian}
\item{\code{P0020034}}{White; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander}
\item{\code{P0020035}}{White; American Indian and Alaska Native; Some Other Race}
\item{\code{P0020036}}{White; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0020037}}{White; Asian; Some Other Race}
\item{\code{P0020038}}{White; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0020039}}{Black or African American; American Indian and Alaska Native; Asian}
\item{\code{P0020040}}{Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander}{P0020021} {Black or African American; Some Other Race}
\item{\code{P0020041}}{Black or African American; American Indian and Alaska Native; Some Other Race}
\item{\code{P0020042}}{Black or African American; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0020043}}{Black or African American; Asian; Some Other Race}
\item{\code{P0020044}}{Black or African American; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0020045}}{American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0020046}}{American Indian and Alaska Native; Asian; Some Other Race}
\item{\code{P0020047}}{American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0020048}}{Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0020049}}{Population of four races;}
\item{\code{P0020050}}{White; Black or African American; American Indian and Alaska Native; Asian}
\item{\code{P0020051}}{White; Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander}
\item{\code{P0020052}}{White; Black or African American; American Indian and Alaska Native; Some Other Race}
\item{\code{P0020053}}{White; Black or African American; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0020054}}{White; Black or African American; Asian; Some Other Race}
\item{\code{P0020055}}{White; Black or African American; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0020056}}{White; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0020057}}{White; American Indian and Alaska Native; Asian; Some Other Race}
\item{\code{P0020058}}{White; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0020059}}{White; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0020060}}{ Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander}{P0020041} {Black or African American; American Indian and Alaska Native; Some Other Race}
\item{\code{P0020061}}{ Black or African American; American Indian and Alaska Native; Asian; Some Other Race}
\item{\code{P0020062}}{Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0020063}}{ Black or African American; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0020064}}{American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0020065}}{Population of five races}
\item{\code{P0020066}}{White; Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0020067}}{White; Black or African American; American Indian and Alaska Native; Asian; Some Other Race}
\item{\code{P0020068}}{White; Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0020069}}{White; Black or African American; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0020070}}{White; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0020071}}{Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0020072}}{Population of six races}
\item{\code{P0020073}}{White; Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0030001}}{Total}
\item{\code{P0030002}}{Population of one race}
\item{\code{P0030003}}{White alone}
\item{\code{P0030004}}{Black or african american alone}
\item{\code{P0030005}}{American Indian and Alaska native alone}
\item{\code{P0030006}}{Asian Alone}
\item{\code{P0030007}}{Native Hawaiian and pacific islander alone}
\item{\code{P0030008}}{Some other race alone}
\item{\code{P0030009}}{Population of two or more races}
\item{\code{P0030010}}{Population of two races}
\item{\code{P0030011}}{White; Black or African American}
\item{\code{P0030012}}{White; American Indian and Alaska Native}
\item{\code{P0030013}}{White; Asian}
\item{\code{P0030014}}{White; Native Hawaiian and Other Pacific Islander}
\item{\code{P0030015}}{White; Some other race}
\item{\code{P0030016}}{Black or African American; American Indian and Alaska Native}
\item{\code{P0030017}}{Black or African American; Asian}
\item{\code{P0030018}}{Black or African American; Native Hawaiian and Other Pacific Islander}
\item{\code{P0030019}}{Black or African American; Some other race}
\item{\code{P0030020}}{American Indian and Alaska Native; Asian}
\item{\code{P0030021}}{American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander}
\item{\code{P0030022}}{American Indian and Alaska Native; Some other race}
\item{\code{P0030023}}{Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0030024}}{Asian; Some other race}
\item{\code{P0030025}}{Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0030026}}{Population of three races}
\item{\code{P0030027}}{White; Black or African American; American Indian and Alaska Native}
\item{\code{P0030028}}{White; Black or African American; Asian}
\item{\code{P0030029}}{White; Black or African American; Native Hawaiian and Other Pacific Islander}
\item{\code{P0030030}}{White; Black or African American; Some other race}
\item{\code{P0030031}}{White; American Indian and Alaska Native; Asian}
\item{\code{P0030032}}{White; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander}
\item{\code{P0030033}}{White; American Indian and Alaska Native; Some other race}
\item{\code{P0030034}}{White; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0030035}}{White; Asian; Some other race}
\item{\code{P0030036}}{White; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0030037}}{Black or African American; American Indian and Alaska Native; Asian}
\item{\code{P0030038}}{Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander}
\item{\code{P0030039}}{Black or African American; American Indian and Alaska Native; Some other race}
\item{\code{P0030040}}{Black or African American; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0030041}}{Black or African American; Asian; Some other race}
\item{\code{P0030042}}{Black or African American; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0030043}}{American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0030044}}{American Indian and Alaska Native; Asian; Some other race}
\item{\code{P0030045}}{American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0030046}}{Asian; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0030047}}{Population of four races}
\item{\code{P0030048}}{White; Black or African American; American Indian and Alaska Native; Asian}
\item{\code{P0030049}}{White; Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander}
\item{\code{P0030050}}{White; Black or African American; American Indian and Alaska Native; Some other race}
\item{\code{P0030051}}{White; Black or African American; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0030052}}{White; Black or African American; Asian; Some other race}
\item{\code{P0030053}}{White; Black or African American; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0030054}}{White; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0030055}}{White; American Indian and Alaska Native; Asian; Some other race}
\item{\code{P0030056}}{White; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0030057}}{White; Asian; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0030058}}{Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0030059}}{ Black or African American; American Indian and Alaska Native; Asian; Some other race}
\item{\code{P0030060}}{Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0030061}}{Black or African American; Asian; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0030062}}{American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0030063}}{Population of five races}
\item{\code{P0030064}}{White; Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0030065}}{White; Black or African American; American Indian and Alaska Native; Asian; Some other race}
\item{\code{P0030066}}{ White; Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0030067}}{White; Black or African American; Asian; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0030068}}{White; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0030069}}{Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0030070}}{Population of six races}
\item{\code{P0030071}}{White; Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0040001}}{Total}
\item{\code{P0040002}}{Hispanic or latino}
\item{\code{P0040003}}{Not hispanic or latino}
\item{\code{P0040004}}{Population of one race}
\item{\code{P0040005}}{White alone}
\item{\code{P0040006}}{Black or african american alone}
\item{\code{P0040007}}{American Indian and alaska native alone}
\item{\code{P0040008}}{Asian Alone}
\item{\code{P0040009}}{Native Hawaiian and pacific islander alone}
\item{\code{P0040010}}{Some other race alone}
\item{\code{P0040011}}{Population of two or more races}
\item{\code{P0040012}}{Population of two races}
\item{\code{P0040013}}{White; Black or African American}
\item{\code{P0040014}}{White; American Indian and Alaska Native}
\item{\code{P0040015}}{White; Asian}
\item{\code{P0040016}}{White; Native Hawaiian and Other Pacific Islander}
\item{\code{P0040017}}{White; Some other race}
\item{\code{P0040018}}{Black or African American; American Indian and Alaska Native}
\item{\code{P0040019}}{Black or African American; Asian}
\item{\code{P0040020}}{ Black or African American; Native Hawaiian and Other Pacific Islander}
\item{\code{P0040021}}{Black or African American; Some other race}
\item{\code{P0040022}}{American Indian and Alaska Native; Asian}
\item{\code{P0040023}}{American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander}
\item{\code{P0040024}}{American Indian and Alaska Native; Some other race}
\item{\code{P0040025}}{Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0040026}}{Asian; SOme other race}
\item{\code{P0040027}}{Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0040028}}{Population of three races}
\item{\code{P0040029}}{White; Black or African American; American Indian and Alaska Native}
\item{\code{P0040030}}{White; Black or African American; Asian}
\item{\code{P0040031}}{White; Black or African American; Native Hawaiian and Other Pacific Islander}
\item{\code{P0040032}}{White; Black or African American; Some other race}
\item{\code{P0040033}}{White; American Indian and Alaska Native; Asian}
\item{\code{P0040034}}{White; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander}
\item{\code{P0040035}}{White; American Indian and Alaska Native; Some other race}
\item{\code{P0040036}}{White; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0040037}}{White; Asian; Some other race}
\item{\code{P0040038}}{White; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0040039}}{Black or African American; American Indian and Alaska Native; Asian}
\item{\code{P0040040}}{Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander}
\item{\code{P0040041}}{Black or African American; American Indian and Alaska Native; Some other race}
\item{\code{P0040042}}{Black or African American; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0040043}}{Black or african american; Asian; some other race}
\item{\code{P0040044}}{Black or African American; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0040045}}{American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0040046}}{American Indian and Alaska Native; Asian; Some other race}
\item{\code{P0040047}}{American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0040048}}{Asian; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0040049}}{Population of four races}
\item{\code{P0040050}}{White; Black or African American; American Indian and Alaska Native; Asian}
\item{\code{P0040051}}{White; Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander}
\item{\code{P0040052}}{White; Black or African American; American Indian and Alaska Native; Some other race}
\item{\code{P0040053}}{White; Black or African American; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0040054}}{White; Black or African American; Asian; Some other race}
\item{\code{P0040055}}{White; Black or African American; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0040056}}{White; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0040057}}{White; American Indian and Alaska Native; Asian; Some other race}
\item{\code{P0040058}}{White; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0040059}}{White; Asian; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0040060}}{Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0040061}}{Black or African American; American Indian and Alaska Native; Asian; Some other race}
\item{\code{P0040062}}{Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0040063}}{Black or African American; Asian; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0040064}}{American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0040065}}{Population of five races}
\item{\code{P0040066}}{White; Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0040067}}{White; Black or African American; American Indian and Alaska Native; Asian; Some other race}
\item{\code{P0040068}}{White; Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0040069}}{White; Black or African American; Asian; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0040070}}{White; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0040071}}{Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0040072}}{Popukation of six races}
\item{\code{P0040073}}{White; Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{H0010001}}{Total}
\item{\code{H0010002}}{Occupied}
\item{\code{H0010003}}{Vacant}
\item{\code{P0050001}}{Total}
\item{\code{P0050002}}{Instituonalized population}
\item{\code{P0050003}}{Correctional facilities for adults}
\item{\code{P0050004}}{Juvenile facilities}
\item{\code{P0050005}}{Nursing facilities/Skilled-nursing facilities}
\item{\code{P0050006}}{Other institutional facilities}
\item{\code{P0050007}}{Noninstitutionalized population:}
\item{\code{P0050008}}{College/University student housing}
\item{\code{P0050009}}{Military quarters}
\item{\code{P0050010}}{Other noninstitutional facilities}
\item{\code{MTFCC20}}{MAT/TIGER Feature Class Code}
\item{\code{geometry}}{Geometric data for the shape. (Point, Multipoint, Polyline, Polygon,etc)}
}}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 67 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 30 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 15 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 75 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 58 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 64 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 8 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 3 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 67 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 159 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 5 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 44 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 102 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 92 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 99 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 105 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 120 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 64 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 16 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 24 rows and 400 columns.

An object of class \code{data.frame} with 14 rows and 398 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 83 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 87 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 82 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 115 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 56 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 93 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 17 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 10 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 21 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 33 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 62 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 100 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 53 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 88 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 77 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 36 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 67 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 78 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 5 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 46 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 66 rows and 400 columns.

An object of class \code{data.frame} with 95 rows and 398 columns.

An object of class \code{data.frame} with 254 rows and 398 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 29 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 14 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 133 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 39 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 55 rows and 400 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 72 rows and 400 columns.

An object of class \code{data.frame} with 23 rows and 398 columns.
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
alabamacounty20

alaskacounty20

arizonacounty20

arkansascounty20

californiacounty20

coloradocounty20

connecticutcounty20

delawarecounty20

dccounty20

floridacounty20

georgiacounty20

hawaiicounty20

idahocounty20

illinoiscounty20

indianacounty20

iowacounty20

kansascounty20

kentuckycounty20

louisianacounty20

mainecounty20

marylandcounty20

massachussetscounty20

michigancounty20

minnesotacounty20

mississippicounty20

missouricounty20

montanacounty20

nebraskacounty20

nevadacounty20

newhampshirecounty20

newjerseycounty20

newmexicocounty20

newyorkcounty20

northcarolinacounty20

northdakotacounty20

ohiocounty20

oklahomacounty20

oregoncounty20

pennsylvaniacounty20

puertoricocounty20

rhodeislandcounty20

southcarolinacounty20

southdakotacounty20

tennessecounty20

texascounty20

utahcounty20

vermontcounty20

virginiacounty20

washingtoncounty20

westvirginiacounty20

wisconsincounty20

wyomingcounty20
}
\description{
UScensuscounty20
This package contains county level data for the US census 2020 redistricting files and shape files.
\describe{
\item{\code{FILEID}}{File Identification}
\item{\code{STUSAB}}{State/US-Abbreviation (USPS)}
\item{\code{SUMLEV}}{Summary Level=050 for county}
\item{\code{GEOVAR}}{Geographic Variant}
\item{\code{GEOCOMP}}{Geographic Component}
\item{\code{CHARITER}}{Characteristic Iteration}
\item{\code{CIFSN}}{Characteristic Iteration File Sequence Number}
\item{\code{LOGRECNO}}{Logical Record Number}
\item{\code{GEOID}}{Geographic Record Identifier}
\item{\code{GEOCODE}}{Geographic Code Identifier}
\item{\code{REGION}}{Region}
\item{\code{DIVISION}}{Division}
\item{\code{STATE}}{State}
\item{\code{COUNTY}}{County}
\item{\code{COUNTYCC}}{FIPS County class code}
\item{\code{COUNTYNS}}{County (NS)}
\item{\code{COUSUB}}{County Sub Divison (FIPS)}
\item{\code{COUSUBCC}}{FIPS County Subdivision Class Code}
\item{\code{COUSUBNS}}{County Subdivision (NS)}
\item{\code{SUBMCD}}{Subminor Civil Division (FIPS)}
\item{\code{SUBMCDCC}}{FIPS Subminor Civil Division Class Code}
\item{\code{SUBMCDNS}}{Subminor Civil Division (NS)}
\item{\code{ESTATE}}{Estate (FIPS)}
\item{\code{ESTATECC}}{FIPS Estate Class Code}
\item{\code{ESTATENS}}{Estate (NS)}
\item{\code{CONCIT}}{Consolidated City (FIPS)}
\item{\code{CONCITCC}}{FIPS Consolidated City Class Code}
\item{\code{CONCITNS}}{Consolidated City (NS)}
\item{\code{PLACE}}{Place (FIPS)}
\item{\code{PLACECC}}{FIPS Place Class Code}
\item{\code{PLACENS}}{Place (NS)}
\item{\code{TRACT}}{Census Tract}
\item{\code{BLKGRP}}{Block Group}
\item{\code{BLOCK}}{Block}
\item{\code{AIANHH}}{American Indian Area/Alaska Native Area/Hawaiian Home Land (Census)}
\item{\code{AIHHTLI}}{American Indian Trust Land/Hawaiian Home Land Indicator}
\item{\code{AIANHHFP}}{American Indian Area/Alaska Native Area/Hawaiian Home Land (FIPS)}
\item{\code{AIANHHCC}}{FIPS American Indian Area/Alaska Native Area/Hawaiian Home Land Class Code}
\item{\code{AIANHHNS}}{American Indian Area/Alaska Native Area/Hawaiian Home Land (NS)}
\item{\code{AITS}}{American Indian Tribal Subdivision (Census)}
\item{\code{AITSFP}}{American Indian Tribal Subdivision (FIPS)}
\item{\code{AITSCC}}{FIPS American Indian Tribal Subdivision Class Code}
\item{\code{AITSNS}}{American Indian Tribal Subdivision (NS)}
\item{\code{TTRACT}}{Tribal Census Tract}
\item{\code{TBLKGRP}}{Tribal Block Group}
\item{\code{ANRC}}{Alaska Native Regional Corporation (FIPS)}
\item{\code{ANRCCC}}{FIPS Alaska Native Regional Corporation Class Code}
\item{\code{ANRCNS}}{Alaska Native Regional Corporation (NS)}
\item{\code{CBSA}}{Metropolitan Statistical Area/Micropolitan Statistical Area}
\item{\code{MEMI}}{Metropolitan/Micropolitan Indicator}
\item{\code{CSA}}{Combine Statistical Area}
\item{\code{METDIV}}{Metropolitan Division}
\item{\code{NECTA}}{New England City and Town Area}
\item{\code{NMEMI}}{NECTA Metropolitan/Micropolitan Indicator}
\item{\code{CNECTA}}{Combined New England City and Town Area}
\item{\code{NECTADIV}}{New England city and town division}
\item{\code{CBSAPCI}}{Metropolitan Statistical Area/Micropolitan Statistical Area Principal City Indicator}
\item{\code{NECTAPCI}}{New England City and Town Area Principal City Indicator}
\item{\code{UA}}{Urban Area}
\item{\code{UATYPE}}{Urban Area Type}
\item{\code{UR}}{Urban/Rural}
\item{\code{CD116}}{Congressional District (116th)}
\item{\code{CD118}}{Congressional District (118th)}
\item{\code{CD119}}{Congressional District (119th)}
\item{\code{CD120}}{Congressional District (120th)}
\item{\code{CD121}}{Congressional District (121th)}
\item{\code{SLDU18}}{State Legislative District (Upper Chamber) (2018)}
\item{\code{SLDU22}}{State Legislative District (Upper Chamber) (2022)}
\item{\code{SLDU24}}{State Legislative District (Upper Chamber) (2024)}
\item{\code{SLDU26}}{State Legislative District (Upper Chamber) (2026)}
\item{\code{SLDU28}}{State Legislative District (Upper Chamber) (2028)}
\item{\code{SLDL18}}{State Legislative District (Lower Chamber) (2018)}
\item{\code{SLDL22}}{State Legislative District (Lower Chamber) (2022)}
\item{\code{SLDL24}}{State Legislative District (Lower Chamber) (2024)}
\item{\code{SLDL26}}{State Legislative District (Lower Chamber) (2026)}
\item{\code{SLDL28}}{State Legislative District (Lower Chamber) (2028)}
\item{\code{VTD}}{Voting District}
\item{\code{VTDI}}{Voting District Indicator}
\item{\code{ZCTA}}{ZIP Code Tabulation Area (5-Digit)}
\item{\code{SDELM}}{School District (Elementary)}
\item{\code{SDSEC}}{School District (Secondary)}
\item{\code{SDUNI}}{School District (Unifield)}
\item{\code{PUMA}}{Public Use Microdata Area}
\item{\code{AREALAND}}{Area(Land)}
\item{\code{AREAWATR}}{Area(Water)}
\item{\code{BASENAME}}{Area Base Name}
\item{\code{NAME}}{Area Name-Legal/Statistical Area Description (LSAD) Term-Part Indicator}
\item{\code{FUNCSTAT}}{Functional Status Code}
\item{\code{GCUNI}}{Geographic Change User Note Indicator}
\item{\code{POP100}}{Population Count (100\%)}
\item{\code{HU100}}{Housing Unit Count (100\%)}
\item{\code{INTPTLAT}}{Internal Point (Latitude)}
\item{\code{INTPTLON}}{Internal Point (Longitude)}
\item{\code{LSADC}}{Legal/Statistical Area Description Code}
\item{\code{PARTFLAG}}{Part Flag}
\item{\code{UGA}}{Urban Growth Area}
\item{\code{P0010001}}{Total Population}
\item{\code{P0010002}}{Population of single race}
\item{\code{P0010003}}{Population of white race alone}
\item{\code{P0010004}}{Population of Black or African american population alone}
\item{\code{P0010005}}{American Indian and Alaskan native alone}
\item{\code{P0010006}}{Asian Alone}
\item{\code{P0010007}}{Native Hawaiian and Other Pacific Islander alone}
\item{\code{P0010008}}{Some other race alone}
\item{\code{P0010009}}{Population of two or more races}
\item{\code{P0010010}}{Population of two races}
\item{\code{P0010011}}{White and black or african american population}
\item{\code{P0010012}}{White and American Indian or alaskan native population}
\item{\code{P0010013}}{White and Asian population}
\item{\code{P0010014}}{White and Native Hawaiian and other Pacific Islander population}
\item{\code{P0010015}}{White and some other race}
\item{\code{P0010016}}{Black or African American and American Indian and Alaska Native Population}
\item{\code{P0010017}}{Black or African American and Asian Population}
\item{\code{P0010018}}{Black or African American and Native Hawaiian and other Pacific Islander Population}
\item{\code{P0010019}}{Black or African American and some other race population}
\item{\code{P0010020}}{American Indian and Alaska Native and Asian Population}
\item{\code{P0010021}}{American Indian and Alaska Native and Native Hawaiian and other pacific islander population}
\item{\code{P0010022}}{American Indian and Alaska Native and some other race population}
\item{\code{P0010023}}{Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0010024}}{Asian and some other race population}
\item{\code{P0010025}}{Native Hawaiian and Other Pacific Islander; Some Other Race population }
\item{\code{P0010026}}{Population of three races}
\item{\code{P0010027}}{white; Black or African American; American Indian and Alaska Native population}
\item{\code{P0010028}}{White; Black or african american; Asian Population}
\item{\code{P0010029}}{white; Black or African American; Native Hawaiian and other pacific islander population}
\item{\code{P0010030}}{White; Black or African American; Some Other Race population}
\item{\code{P0010031}}{White;  American Indian and Alaska Native; Asian population }
\item{\code{P0010032}}{White; American Indian and Alaska Native; Native Hawaiian and other pacific islander population}
\item{\code{P0010033}}{White; American Indian and Alaska Native;  Some other race population}
\item{\code{P0010034}}{White; Asian; Native Hawaiian and other pacific Islander Population}
\item{\code{P0010035}}{White; Asian; Some other race population}
\item{\code{P0010036}}{White; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0010037}}{Black or African American; American Indian and Alaska Native; Asian Population}
\item{\code{P0010038}}{Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander population}
\item{\code{P0010039}}{Black or African American; American Indian and Alaska Native; Some Other Race population}
\item{\code{P0010040}}{Black or African American; Asian; Native Hawaiian and Other Pacific Islander Population}
\item{\code{P0010041}}{Black or African American; Asian; Some Other Race population}
\item{\code{P0010042}}{Black or African American; Native Hawaiian and Other Pacific Islander; Some Other Race population}
\item{\code{P0010043}}{American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander Population}
\item{\code{P0010044}}{American Indian and Alaska Native; Asian; Some Other Race population}
\item{\code{P0010045}}{American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race population}
\item{\code{P0010046}}{Asian; Native Hawaiian and Other Pacific Islander; Some Other Race population}
\item{\code{P0010046}}{Asian; Native Hawaiian and Other Pacific Islander; Some Other Race population}
\item{\code{P0010047}}{Population of four races}
\item{\code{P0010048}}{White; Black or African American; American Indian and Alaska Native; Asian}
\item{\code{P0010049}}{White; Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander}
\item{\code{P0010050}}{White; Black or African American; American Indian and Alaska Native; Some Other Race}
\item{\code{P0010051}}{White; Black or African American; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0010052}}{White; Black or African American; Asian; Some Other Race}
\item{\code{P0010053}}{White; Black or African American; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0010054}}{White; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0010055}}{White; American Indian and Alaska Native; Asian; Some Other Race}
\item{\code{P0010056}}{White; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0010057}}{White; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0010058}}{Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0010059}}{Black or African American; American Indian and Alaska Native; Asian; Some Other Race}
\item{\code{P0010060}}{Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0010061}}{Black or African American; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0010062}}{American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0010063}}{Population of five races}
\item{\code{P0010064}}{White; Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0010065}}{White; Black or African American; American Indian and Alaska Native; Asian; Some Other Race}
\item{\code{P0010066}}{White; Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0010067}}{White; Black or African American; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0010068}}{White; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0010069}}{Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0010070}}{Population of six races}
\item{\code{P0010071}}{White; Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0020001}}{Total:}
\item{\code{P0020002}}{Hispanic or Latino}
\item{\code{P0020003}}{Not Hispanic or Latino}
\item{\code{P0020004}}{Population of one race}
\item{\code{P0020005}}{White alone}
\item{\code{P0020006}}{Black or african american alone}
\item{\code{P0020007}}{American Indian or Alaskan Native alone}
\item{\code{P0020008}}{Asian Alone}
\item{\code{P0020009}}{Native Hawaiian and Other Pacific Islander alone}
\item{\code{P0020010}}{Some other race alone}
\item{\code{P0020011}}{Population of two or more races}
\item{\code{P0020012}}{Population of two races}
\item{\code{P0020013}}{White; Black or African American}
\item{\code{P0020014}}{White; American Indian and Alaska Native}
\item{\code{P0020015}}{White; Asian}
\item{\code{P0020016}}{White; Native Hawaiian and Other Pacific Islander}
\item{\code{P0020017}}{White; Some other race}
\item{\code{P0020018}}{Black or African American; American Indian and Alaska Native}
\item{\code{P0020019}}{Black or African American; Asian}
\item{\code{P0020020}}{Black or African American; Native Hawaiian and Other Pacific Islander}
\item{\code{P0020021}}{Black or African American; Some Other Race}
\item{\code{P0020022}}{American Indian and Alaska Native; Asian}
\item{\code{P0020023}}{American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander}
\item{\code{P0020024}}{American Indian and Alaska Native; Some Other Race}
\item{\code{P0020025}}{Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0020026}}{Asian; Some Other Race}
\item{\code{P0020027}}{Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0020028}}{Population of three races}
\item{\code{P0020029}}{White; Black or African American; American Indian and Alaska Native}
\item{\code{P0020030}}{White; Black or African American; Asian}
\item{\code{P0020031}}{White; Black or African American; Native Hawaiian and Other Pacific Islander}
\item{\code{P0020032}}{White; Black or African American; Some Other Race}
\item{\code{P0020033}}{White; American Indian and Alaska Native; Asian}
\item{\code{P0020034}}{White; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander}
\item{\code{P0020035}}{White; American Indian and Alaska Native; Some Other Race}
\item{\code{P0020036}}{White; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0020037}}{White; Asian; Some Other Race}
\item{\code{P0020038}}{White; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0020039}}{Black or African American; American Indian and Alaska Native; Asian}
\item{\code{P0020040}}{Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander}{P0020021} {Black or African American; Some Other Race}
\item{\code{P0020041}}{Black or African American; American Indian and Alaska Native; Some Other Race}
\item{\code{P0020042}}{Black or African American; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0020043}}{Black or African American; Asian; Some Other Race}
\item{\code{P0020044}}{Black or African American; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0020045}}{American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0020046}}{American Indian and Alaska Native; Asian; Some Other Race}
\item{\code{P0020047}}{American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0020048}}{Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0020049}}{Population of four races;}
\item{\code{P0020050}}{White; Black or African American; American Indian and Alaska Native; Asian}
\item{\code{P0020051}}{White; Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander}
\item{\code{P0020052}}{White; Black or African American; American Indian and Alaska Native; Some Other Race}
\item{\code{P0020053}}{White; Black or African American; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0020054}}{White; Black or African American; Asian; Some Other Race}
\item{\code{P0020055}}{White; Black or African American; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0020056}}{White; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0020057}}{White; American Indian and Alaska Native; Asian; Some Other Race}
\item{\code{P0020058}}{White; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0020059}}{White; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0020060}}{ Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander}{P0020041} {Black or African American; American Indian and Alaska Native; Some Other Race}
\item{\code{P0020061}}{ Black or African American; American Indian and Alaska Native; Asian; Some Other Race}
\item{\code{P0020062}}{Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0020063}}{ Black or African American; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0020064}}{American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0020065}}{Population of five races}
\item{\code{P0020066}}{White; Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0020067}}{White; Black or African American; American Indian and Alaska Native; Asian; Some Other Race}
\item{\code{P0020068}}{White; Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0020069}}{White; Black or African American; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0020070}}{White; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0020071}}{Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0020072}}{Population of six races}
\item{\code{P0020073}}{White; Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
\item{\code{P0030001}}{Total}
\item{\code{P0030002}}{Population of one race}
\item{\code{P0030003}}{White alone}
\item{\code{P0030004}}{Black or african american alone}
\item{\code{P0030005}}{American Indian and Alaska native alone}
\item{\code{P0030006}}{Asian Alone}
\item{\code{P0030007}}{Native Hawaiian and pacific islander alone}
\item{\code{P0030008}}{Some other race alone}
\item{\code{P0030009}}{Population of two or more races}
\item{\code{P0030010}}{Population of two races}
\item{\code{P0030011}}{White; Black or African American}
\item{\code{P0030012}}{White; American Indian and Alaska Native}
\item{\code{P0030013}}{White; Asian}
\item{\code{P0030014}}{White; Native Hawaiian and Other Pacific Islander}
\item{\code{P0030015}}{White; Some other race}
\item{\code{P0030016}}{Black or African American; American Indian and Alaska Native}
\item{\code{P0030017}}{Black or African American; Asian}
\item{\code{P0030018}}{Black or African American; Native Hawaiian and Other Pacific Islander}
\item{\code{P0030019}}{Black or African American; Some other race}
\item{\code{P0030020}}{American Indian and Alaska Native; Asian}
\item{\code{P0030021}}{American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander}
\item{\code{P0030022}}{American Indian and Alaska Native; Some other race}
\item{\code{P0030023}}{Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0030024}}{Asian; Some other race}
\item{\code{P0030025}}{Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0030026}}{Population of three races}
\item{\code{P0030027}}{White; Black or African American; American Indian and Alaska Native}
\item{\code{P0030028}}{White; Black or African American; Asian}
\item{\code{P0030029}}{White; Black or African American; Native Hawaiian and Other Pacific Islander}
\item{\code{P0030030}}{White; Black or African American; Some other race}
\item{\code{P0030031}}{White; American Indian and Alaska Native; Asian}
\item{\code{P0030032}}{White; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander}
\item{\code{P0030033}}{White; American Indian and Alaska Native; Some other race}
\item{\code{P0030034}}{White; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0030035}}{White; Asian; Some other race}
\item{\code{P0030036}}{White; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0030037}}{Black or African American; American Indian and Alaska Native; Asian}
\item{\code{P0030038}}{Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander}
\item{\code{P0030039}}{Black or African American; American Indian and Alaska Native; Some other race}
\item{\code{P0030040}}{Black or African American; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0030041}}{Black or African American; Asian; Some other race}
\item{\code{P0030042}}{Black or African American; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0030043}}{American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0030044}}{American Indian and Alaska Native; Asian; Some other race}
\item{\code{P0030045}}{American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0030046}}{Asian; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0030047}}{Population of four races}
\item{\code{P0030048}}{White; Black or African American; American Indian and Alaska Native; Asian}
\item{\code{P0030049}}{White; Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander}
\item{\code{P0030050}}{White; Black or African American; American Indian and Alaska Native; Some other race}
\item{\code{P0030051}}{White; Black or African American; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0030052}}{White; Black or African American; Asian; Some other race}
\item{\code{P0030053}}{White; Black or African American; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0030054}}{White; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0030055}}{White; American Indian and Alaska Native; Asian; Some other race}
\item{\code{P0030056}}{White; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0030057}}{White; Asian; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0030058}}{Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0030059}}{ Black or African American; American Indian and Alaska Native; Asian; Some other race}
\item{\code{P0030060}}{Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0030061}}{Black or African American; Asian; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0030062}}{American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0030063}}{Population of five races}
\item{\code{P0030064}}{White; Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0030065}}{White; Black or African American; American Indian and Alaska Native; Asian; Some other race}
\item{\code{P0030066}}{ White; Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0030067}}{White; Black or African American; Asian; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0030068}}{White; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0030069}}{Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0030070}}{Population of six races}
\item{\code{P0030071}}{White; Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0040001}}{Total}
\item{\code{P0040002}}{Hispanic or latino}
\item{\code{P0040003}}{Not hispanic or latino}
\item{\code{P0040004}}{Population of one race}
\item{\code{P0040005}}{White alone}
\item{\code{P0040006}}{Black or african american alone}
\item{\code{P0040007}}{American Indian and alaska native alone}
\item{\code{P0040008}}{Asian Alone}
\item{\code{P0040009}}{Native Hawaiian and pacific islander alone}
\item{\code{P0040010}}{Some other race alone}
\item{\code{P0040011}}{Population of two or more races}
\item{\code{P0040012}}{Population of two races}
\item{\code{P0040013}}{White; Black or African American}
\item{\code{P0040014}}{White; American Indian and Alaska Native}
\item{\code{P0040015}}{White; Asian}
\item{\code{P0040016}}{White; Native Hawaiian and Other Pacific Islander}
\item{\code{P0040017}}{White; Some other race}
\item{\code{P0040018}}{Black or African American; American Indian and Alaska Native}
\item{\code{P0040019}}{Black or African American; Asian}
\item{\code{P0040020}}{ Black or African American; Native Hawaiian and Other Pacific Islander}
\item{\code{P0040021}}{Black or African American; Some other race}
\item{\code{P0040022}}{American Indian and Alaska Native; Asian}
\item{\code{P0040023}}{American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander}
\item{\code{P0040024}}{American Indian and Alaska Native; Some other race}
\item{\code{P0040025}}{Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0040026}}{Asian; SOme other race}
\item{\code{P0040027}}{Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0040028}}{Population of three races}
\item{\code{P0040029}}{White; Black or African American; American Indian and Alaska Native}
\item{\code{P0040030}}{White; Black or African American; Asian}
\item{\code{P0040031}}{White; Black or African American; Native Hawaiian and Other Pacific Islander}
\item{\code{P0040032}}{White; Black or African American; Some other race}
\item{\code{P0040033}}{White; American Indian and Alaska Native; Asian}
\item{\code{P0040034}}{White; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander}
\item{\code{P0040035}}{White; American Indian and Alaska Native; Some other race}
\item{\code{P0040036}}{White; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0040037}}{White; Asian; Some other race}
\item{\code{P0040038}}{White; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0040039}}{Black or African American; American Indian and Alaska Native; Asian}
\item{\code{P0040040}}{Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander}
\item{\code{P0040041}}{Black or African American; American Indian and Alaska Native; Some other race}
\item{\code{P0040042}}{Black or African American; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0040043}}{Black or african american; Asian; some other race}
\item{\code{P0040044}}{Black or African American; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0040045}}{American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0040046}}{American Indian and Alaska Native; Asian; Some other race}
\item{\code{P0040047}}{American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0040048}}{Asian; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0040049}}{Population of four races}
\item{\code{P0040050}}{White; Black or African American; American Indian and Alaska Native; Asian}
\item{\code{P0040051}}{White; Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander}
\item{\code{P0040052}}{White; Black or African American; American Indian and Alaska Native; Some other race}
\item{\code{P0040053}}{White; Black or African American; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0040054}}{White; Black or African American; Asian; Some other race}
\item{\code{P0040055}}{White; Black or African American; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0040056}}{White; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0040057}}{White; American Indian and Alaska Native; Asian; Some other race}
\item{\code{P0040058}}{White; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0040059}}{White; Asian; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0040060}}{Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0040061}}{Black or African American; American Indian and Alaska Native; Asian; Some other race}
\item{\code{P0040062}}{Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0040063}}{Black or African American; Asian; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0040064}}{American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0040065}}{Population of five races}
\item{\code{P0040066}}{White; Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander}
\item{\code{P0040067}}{White; Black or African American; American Indian and Alaska Native; Asian; Some other race}
\item{\code{P0040068}}{White; Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0040069}}{White; Black or African American; Asian; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0040070}}{White; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0040071}}{Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{P0040072}}{Popukation of six races}
\item{\code{P0040073}}{White; Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some other race}
\item{\code{H0010001}}{Total}
\item{\code{H0010002}}{Occupied}
\item{\code{H0010003}}{Vacant}
\item{\code{P0050001}}{Total}
\item{\code{P0050002}}{Instituonalized population}
\item{\code{P0050003}}{Correctional facilities for adults}
\item{\code{P0050004}}{Juvenile facilities}
\item{\code{P0050005}}{Nursing facilities/Skilled-nursing facilities}
\item{\code{P0050006}}{Other institutional facilities}
\item{\code{P0050007}}{Noninstitutionalized population:}
\item{\code{P0050008}}{College/University student housing}
\item{\code{P0050009}}{Military quarters}
\item{\code{P0050010}}{Other noninstitutional facilities}
\item{\code{MTFCC20}}{MAT/TIGER Feature Class Code}
\item{\code{geometry}}{Geometric data for the shape. (Point, Multipoint, Polyline, Polygon,etc)}
}
}
\keyword{datasets}
